Fold BlockNumberStreamingProtocolSchedule into MilestoneStreamingProtocolSchedule Signed-off-by: Simon Dudley <simon.dudley@consensys.net>
